Try to fix Mac OS X buildbot packet-http.c:2629: warning: implicit conversion shortens 64-bit value into a 32-bit value
Reviewed-on: https://code.wireshark.org/review/1389

+++ epan/dissectors/packet-http.c
@@ -2626,7 +2626,7 @@ process_header(tvbuff_t *tvb, int offset, int next_offset,
 					part = value + i;
 					part_end = (char *)memchr(part, ';', value_len - i);
 					if (part_end)
-						part_len = part_end - part;
+						part_len =(int)(part_end - part);
 					else
 						part_len = value_len - i;
